pyproject.toml
[project]
name = "ai-biotech-ml-portfolio"
version = "0.1.0"
description = "Streamlit demos for AI Biotech & Translational ML"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.36,<2",
    "scikit-learn>=1.3,<2",
    "pandas>=2.1,<3",
    "numpy>=1.24,<3",
    "plotly>=5.20,<6",
    "shap>=0.41,<1",
    "fastapi>=0.110,<1",
    "uvicorn>=0.23,<1",
    "pydantic>=2.7,<3",
    "umap-learn>=0.5,<1",
    "scipy>=1.10,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<9",
    "httpx>=0.27,<1",
]

[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"


